requirements.txt
customtkinter
Pillow
mutagen
fastapi
uvicorn
pygame
websockets